# gh-104144: Optimize gather to finish eagerly when all futures complete eagerly