--- a/Misc/NEWS.d/next/Library/2023-05-03-16-50-24.gh-issue-104144.yNkjL8.rst
+++ b/Misc/NEWS.d/next/Library/2023-05-03-16-50-24.gh-issue-104144.yNkjL8.rst
@@ -1,3 +1,3 @@
-Optimize asyncio.gather when using eager tasks factory Skip creating
-gathering future and scheduling done callbacks when all futures finish
-without blocking - for up to 3x speedup
+Optimize :func:`asyncio.gather` when using :func:`asyncio.eager_task_factory`.
+Skip creating gathering future and scheduling done callbacks when all futures finish
+without blocking.