--- a/Misc/NEWS.d/next/Library/2023-05-03-16-50-24.gh-issue-104144.yNkjL8.rst
+++ b/Misc/NEWS.d/next/Library/2023-05-03-16-50-24.gh-issue-104144.yNkjL8.rst
@@ -1,3 +1,3 @@
-Optimize :func:`asyncio.gather` when using :func:`asyncio.eager_task_factory`.
-Skip creating gathering future and scheduling done callbacks when all futures finish
-without blocking.
+Optimize :func:`asyncio.gather` when using :func:`asyncio.eager_task_factory`
+to complete eagerly if all fututres completed eagerly.
+Avoid scheduling done callbacks for fututres that complete eagerly.