@itamaro 2023-05-05
thanks for the review @jbower-fb !

I pushed a new version of the PR based on your suggestions, but not identical.
summary of my changes:
- instead of immediately calling the done callbacks for eagerly completed futures in the loop, I add them to a list and call their callbacks only at the end, after creating the outer future
- this helps because now outer is defined the same way, and the children list is fully populated, which is also important because the done callback uses the children list
- I always create a GatheringFuture so I get the same cancellation treatment, but it will still finish eagerly if all futures completed eagerly thanks to the last done callback right before returning
- there shouldn't be any issues with bpo-46672 since outer looks the same as before