@itamaro 2023-05-05
> The title no longer accurately describes the updated PR, since the `_GatheringFuture` is always created.

thanks, I updated the title!