@gvanrossum 2023-05-06
PS. In _general_ there's no need to click the "Update branch" button (or otherwise merge main back into the PR) *unless* there are fixes/changes that might affect the PR (e.g. if touching the same file).